refactor(ptv3): move SemsegModule to its own file
Relocate the SemsegModule class from ptv3_trt.{hpp,cpp} into
semseg_module.{hpp,cpp}. Pure move plus the include and CMake source entry.
