+++ perception/autoware_ptv3/include/autoware/ptv3/ptv3_trt.hpp
@@ -17,6 +17,7 @@
 
 #include "autoware/ptv3/backbone_preprocessor.hpp"
 #include "autoware/ptv3/backbone_engine.hpp"
+#include "autoware/ptv3/semseg_module.hpp"
 #include "autoware/ptv3/execution_context.hpp"
 #include "autoware/ptv3/postprocess/detection3d_postprocess.hpp"
 #include "autoware/ptv3/postprocess/postprocess_kernel.hpp"
@@ -45,86 +46,6 @@
 
 namespace autoware::ptv3
 {
-
-class PTV3_PUBLIC SemsegModule
-{
-public:
-  SemsegModule(
-    const tensorrt_common::TrtCommonConfig & trt_config, const PTv3SemsegConfig & config,
-    const float * backbone_point_feat, const float * backbone_input_features,
-    const void * compact_points);
-
-  [[nodiscard]] bool shouldRun(
-    bool should_publish_segmented_pointcloud, bool should_publish_visualization_pointcloud,
-    bool should_publish_filtered_pointcloud) const;
-
-  void setPublishSegmentedPointcloud(
-    std::function<void(std::unique_ptr<const cuda_blackboard::CudaPointCloud2>)> func);
-  void setPublishVisualizationPointcloud(
-    std::function<void(std::unique_ptr<const cuda_blackboard::CudaPointCloud2>)> func);
-  void setPublishFilteredPointcloud(
-    std::function<void(std::unique_ptr<const cuda_blackboard::CudaPointCloud2>)> func);
-
-  void preparePreprocess(
-    const cuda_blackboard::CudaPointCloud2 & msg, const PTv3ExecutionContext & context,
-    CloudFormat input_format, bool should_run);
-  void preprocessSourceReconstruction(
-    const cuda_blackboard::CudaPointCloud2 & msg, const PTv3ExecutionContext & context,
-    const PreprocessCuda & backbone_preprocessor, std::size_t num_cropped_points, bool should_run);
-  void setInputShape(std::int64_t num_voxels);
-  bool enqueue(const PTv3ExecutionContext & context);
-  bool postProcess(
-    const std_msgs::msg::Header & header, const PTv3ExecutionContext & context,
-    const PreprocessCuda & preprocessor, CloudFormat input_format,
-    bool should_publish_segmented_pointcloud, bool should_publish_visualization_pointcloud,
-    bool should_publish_filtered_pointcloud, std::int64_t num_voxels);
-
-private:
-  void initTrt(const tensorrt_common::TrtCommonConfig & trt_config);
-  void createPointFields();
-  void prepareFilteredPointcloudFormat(CloudFormat input_format);
-  void allocateOutputMessages();
-  [[nodiscard]] std::int64_t outputCapacity() const;
-
-  PTv3SemsegConfig config_;
-  const float * backbone_point_feat_;
-  const float * backbone_input_features_;
-  const void * compact_points_;
-
-  std::unique_ptr<autoware::tensorrt_common::TrtCommon> trt_ptr_{nullptr};
-  std::unique_ptr<SemsegPreprocessCuda> preprocess_ptr_{nullptr};
-  std::unique_ptr<PostprocessCuda> post_ptr_{nullptr};
-
-  std::function<void(std::unique_ptr<const cuda_blackboard::CudaPointCloud2>)>
-    publish_segmented_pointcloud_{nullptr};
-  std::function<void(std::unique_ptr<const cuda_blackboard::CudaPointCloud2>)>
-    publish_visualization_pointcloud_{nullptr};
-  std::function<void(std::unique_ptr<const cuda_blackboard::CudaPointCloud2>)>
-    publish_filtered_pointcloud_{nullptr};
-
-  std::vector<sensor_msgs::msg::PointField> segmented_pointcloud_fields_;
-  std::vector<sensor_msgs::msg::PointField> visualization_pointcloud_fields_;
-  std::vector<sensor_msgs::msg::PointField> filtered_pointcloud_fields_;
-
-  std::unique_ptr<cuda_blackboard::CudaPointCloud2> segmented_points_msg_ptr_{nullptr};
-  std::unique_ptr<cuda_blackboard::CudaPointCloud2> visualization_points_msg_ptr_{nullptr};
-  std::unique_ptr<cuda_blackboard::CudaPointCloud2> filtered_points_msg_ptr_{nullptr};
-
-  std::size_t num_source_points_{0};
-  std::size_t num_cropped_points_{0};
-  const void * current_input_data_{nullptr};
-  std::once_flag init_filtered_cloud_;
-  CloudFormat input_format_{CloudFormat::UNKNOWN};
-  CloudFormat filtered_output_format_{CloudFormat::UNKNOWN};
-
-  autoware::cuda_utils::CudaUniquePtr<std::uint8_t[]> cropped_source_points_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<float[]> reconstructed_features_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> inverse_map_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> reconstructed_labels_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<float[]> reconstructed_probs_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pred_labels_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<float[]> pred_probs_d_{nullptr};
-};
 
 class PTV3_PUBLIC Detection3DModule
 {
